app/requirements.txt
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
psycopg2-binary==2.9.9
plotly==5.18.0
altair==5.2.0
scikit-learn==1.3.2
prophet==1.1.5
python-dotenv==1.0.0
sqlalchemy==2.0.23